Refactor HeroSection component to move code to separate folder and update imports
